# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Primavera V10 API - LLM Ready",
    "version": "4.0.0",
    "description": "API REST otimizada para consumo por LLMs. Dados do ERP Cegid Primavera V10 da Cimenteira do Louro. Todos os cálculos de faturação usam 34 tipos de documento com notas de crédito como valores negativos.",
    "contact": {
      "name": "Cimenteira do Louro",
      "url": "http://primavera-api.cimenteiradolouro.local"
    }
  },
  "servers": [
    {
      "url": "http://primavera-api.cimenteiradolouro.local",
      "description": "Servidor Primavera API"
    }
  ],
  "security": [
    { "basicAuth": [] },
    { "apiKeyHeader": [] },
    { "apiKeyQuery": [] }
  ],
...
  "paths": {
...
    "/api/comercial/resumo": {
      "get": {
        "tags": ["Comercial"],
        "summary": "Resumo comercial",
        "description": "Resumo de vendas do mês e ano corrente com comparações ao período anterior.",
        "operationId": "resumoComercial",
        "parameters": [
          {
            "name": "vendedor",
            "in": "query",
            "description": "Filtrar por código do vendedor",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Resumo de vendas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "mesAtual": {
                          "type": "object",
                          "properties": {
                            "faturacao": { "type": "number", "format": "double", "example": 245000.00 },
                            "numDocumentos": { "type": "integer", "example": 128 },
                            "numClientes": { "type": "integer", "example": 45 },
                            "ticketMedio": { "type": "number", "format": "double", "example": 1914.06 }
                          }
                        },
                        "mesAnterior": {
                          "type": "object",
                          "properties": {
                            "faturacao": { "type": "number", "format": "double", "example": 230000.00 },
                            "variacao": { "type": "number", "format": "double", "example": 6.5 }
                          }
                        },
                        "acumuladoAno": {
                          "type": "object",
                          "properties": {
                            "faturacao": { "type": "number", "format": "double", "example": 720000.00 },
                            "faturacaoAnoAnterior": { "type": "number", "format": "double", "example": 680000.00 },
                            "variacao": { "type": "number", "format": "double", "example": 5.9 }
                          }
                        },
                        "mesHomologo": {
                          "type": "object",
                          "properties": {
                            "faturacao": { "type": "number", "format": "double", "example": 225000.00 },
                            "variacao": { "type": "number", "format": "double", "example": 8.9 }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/comercial/por-localidade": {
      "get": {
        "tags": ["Comercial"],
        "summary": "Vendas por localidade",
        "description": "Análise de vendas agrupadas por localidade/zona geográfica.",
        "operationId": "vendasPorLocalidade",
        "parameters": [
          {
            "name": "periodo",
            "in": "query",
            "description": "Período de análise",
            "schema": { "type": "string", "enum": ["mes", "trimestre", "ano"], "default": "ano" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número máximo de localidades",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Vendas por localidade",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "localidade": { "type": "string", "example": "Braga" },
                          "zona": { "type": "string", "example": "Norte" },
                          "faturacao": { "type": "number", "format": "double", "example": 185000.00 },
                          "numClientes": { "type": "integer", "example": 23 },
                          "numDocumentos": { "type": "integer", "example": 156 },
                          "percentagemTotal": { "type": "number", "format": "double", "example": 14.2 }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/comercial/clientes-risco": {
      "get": {
        "tags": ["Comercial"],
        "summary": "Clientes em risco",
        "description": "Identificação de clientes em risco de perda por inatividade ou queda significativa de faturação.",
        "operationId": "clientesRisco",
        "parameters": [
          {
            "name": "diasInativo",
            "in": "query",
            "description": "Número mínimo de dias sem compras para considerar inativo",
            "schema": { "type": "integer", "default": 90 }
          },
          {
            "name": "quedaMinima",
            "in": "query",
            "description": "Percentagem mínima de queda de faturação para considerar risco",
            "schema": { "type": "number", "format": "double", "default": 30 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número máximo de resultados",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes em risco",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "cliente": { "type": "string", "example": "C0045" },
                          "nome": { "type": "string", "example": "Obras Públicas do Minho SA" },
                          "tipoRisco": { "type": "string", "enum": ["inatividade", "queda_faturacao", "ambos"], "example": "queda_faturacao" },
                          "diasSemCompra": { "type": "integer", "example": 45 },
                          "faturacaoAtual": { "type": "number", "format": "double", "example": 12000.00 },
                          "faturacaoAnterior": { "type": "number", "format": "double", "example": 35000.00 },
                          "quedaPercentual": { "type": "number", "format": "double", "example": -65.7 },
                          "ultimaCompra": { "type": "string", "format": "date", "example": "2026-01-24" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/comercial/recomendacoes/{clienteId}": {
      "get": {
        "tags": ["Comercial"],
        "summary": "Recomendações de produtos",
        "description": "Recomendações de produtos para um cliente baseadas no histórico de compras e padrões de clientes similares.",
        "operationId": "recomendacoesProdutos",
        "parameters": [
          {
            "name": "clienteId",
            "in": "path",
            "required": true,
            "description": "Código do cliente",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número máximo de recomendações",
            "schema": { "type": "integer", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de recomendações",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "artigo": { "type": "string", "example": "ADT005" },
                          "descricao": { "type": "string", "example": "Aditivo Plastificante Premium" },
                          "familia": { "type": "string", "example": "Aditivos" },
                          "motivo": { "type": "string", "example": "Comprado por 78% dos clientes similares" },
                          "confianca": { "type": "number", "format": "double", "example": 0.85 },
                          "precoSugerido": { "type": "number", "format": "double", "example": 8.50 }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Autenticação HTTP Basic"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Chave de API via header"
      },
      "apiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey",
        "description": "Chave de API via query parameter"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Não autorizado - credenciais inválidas ou em falta",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string", "example": "Autenticação necessária" },
                "code": { "type": "string", "example": "UNAUTHORIZED" }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Recurso não encontrado",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string", "example": "Recurso não encontrado" },
                "code": { "type": "string", "example": "NOT_FOUND" }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Erro interno do servidor",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string", "example": "Erro interno do servidor" },
                "code": { "type": "string", "example": "INTERNAL_ERROR" }
              }
            }
          }
        }
      }
    }
  }
}